+++ src/index.d.ts
@@ -1,10 +1,5 @@
 // This file contains type declarations for dependencies
 
-declare module 'marked' {
-  function marked(markdownString: string, options?: object, callback?: Function): string;
-  export = marked;
-}
-
 declare module 'seedrandom' {
   // The actual seedrandom declaration is much more complicated than
   // this - but this simplified interface is all we use.